+++ RELEASE_NOTES.md
@@ -5,7 +5,9 @@
 
 - added a `Commands` button to AI chat for quick interaction with the AI agent
 - [openrouter.ai](https://openrouter.ai) can be used as AI agent
+- Model selection added for codex and copilot
 - Smaller MCP responses
 
 ### Bug fixes
 
+- Fixed path finding for copilot cli